feat(hooks): add checkIfExtensionIsDownloaded
