+++ components/core/Upload/Jumper.js
@@ -14,6 +14,7 @@ import { useUploadStore } from "~/components/core/Upload/store";
 import { useUploadOnboardingContext } from "~/components/core/Onboarding/Upload";
 
 import DownloadExtensionButton from "~/components/core/Extension/DownloadExtensionButton";
+import { useCheckIfExtensionIsInstalled } from "~/common/hooks";
 
 const STYLES_EXTENSION_BAR = (theme) => css`
   ${Styles.HORIZONTAL_CONTAINER_CENTERED};
@@ -28,10 +29,12 @@ const STYLES_EXTENSION_BAR = (theme) => css`
 `;
 
 function ExtensionBar() {
+  const { isExtensionDownloaded } = useCheckIfExtensionIsInstalled();
+
   const [isVisible, setVisibility] = React.useState(true);
   const hideExtensionBar = () => setVisibility(false);
 
-  if (!isVisible) return null;
+  if (isExtensionDownloaded || !isVisible) return null;
 
   return (
     <Jumper.Item css={STYLES_EXTENSION_BAR}>
@@ -124,7 +127,7 @@ export function UploadJumper({ data }) {
     setState((prev) => ({ ...prev, [e.target.name]: e.target.value, urlError: false }));
   };
 
-  const isOnboarding = onboardingContext.currentStep === onboardingContext.steps.jumperWalkthrough;
+  const isOnboarding = onboardingContext.currentStep === onboardingContext.steps.jumper;
 
   return (
     <Jumper.AnimatePresence>
@@ -133,8 +136,8 @@ export function UploadJumper({ data }) {
           <Jumper.Header>
             <System.H5 color="textBlack">Upload</System.H5>
           </Jumper.Header>
-          {isOnboarding && <ExtensionBar />}
           <Jumper.Divider />
+          {isOnboarding && <ExtensionBar />}
           <Jumper.Item css={STYLES_LINK_UPLOAD_WRAPPER}>
             <div css={Styles.HORIZONTAL_CONTAINER}>
               <System.Input
